requirements.txt
fastapi
uvicorn
pydantic
pdfminer.six
python-docx
pytesseract
Pillow
cerberus
transformers
gradio
asyncio
